+++ src/main/java/com/chillibits/particulatematterapi/controller/v1/UserController.java
@@ -95,18 +95,6 @@ public class UserController {
     }
 
     /**
-     * Confirms an user account
-     *
-     * @param token Randomly generated confirmation token, which is included in confirmation links
-     * @return Redirect to the pmapp website to show a success / failure message
-     */
-    @RequestMapping(method = RequestMethod.GET, path = "/confirm")
-    @ApiOperation(value = "Confirms an user account", hidden = true)
-    public String confirmAccount(@RequestParam String token) {
-        return "redirect:https://www.chillibits.com/pmapp?p=confirmation&param=" + (userService.confirmAccount(token) ? "success" : "failure");
-    }
-
-    /**
      * Updates an existing user
      * <p>Note: Requires at least application role CBA (ChilliBits application)</p>
      *
